requirements.txt
shazamio==0.5.1
python-telegram-bot==21.0.1
ffmpeg-downloader==0.3.0
SQLAlchemy==2.0.29
pymorphy2==0.9.1